refactor waste runtime handler helpers
